mysite/requirements.txt
Django~=3.2.7
django-crispy-forms
crispy-bootstrap5
cx_Oracle
django-filter